starfish_core

Trait BlockHeaderAPI

pub trait BlockHeaderAPI {

    // Required methods
    fn epoch(&self) -> Epoch;
    fn round(&self) -> Round;
    fn author(&self) -> AuthorityIndex;
    fn slot(&self) -> Slot;
    fn acknowledgments(&self) -> &[BlockRef];
    fn timestamp_ms(&self) -> BlockTimestampMs;
    fn ancestors(&self) -> &[BlockRef];
    fn commit_votes(&self) -> &[CommitRef];
    fn transactions_commitment(&self) -> TransactionsCommitment;
    fn strong_vote(&self) -> Option<StrongVote>;
    fn strong_vote_leader(&self) -> Option<AuthorityIndex>;
    fn is_strong_vote(&self) -> bool;
    fn is_strong_blame(&self) -> bool;

    // Provided methods
    fn is_strong_vote_for(&self, leader: AuthorityIndex) -> bool { ... }
    fn is_strong_blame_for(&self, leader: AuthorityIndex) -> bool { ... }

}

Required Methods

fn epoch(&self) -> Epoch

fn round(&self) -> Round

fn author(&self) -> AuthorityIndex

fn slot(&self) -> Slot

fn acknowledgments(&self) -> &[BlockRef]

fn timestamp_ms(&self) -> BlockTimestampMs

fn ancestors(&self) -> &[BlockRef]

fn commit_votes(&self) -> &[CommitRef]

fn transactions_commitment(&self) -> TransactionsCommitment

fn strong_vote(&self) -> Option<StrongVote>

fn strong_vote_leader(&self) -> Option<AuthorityIndex>

fn is_strong_vote(&self) -> bool

fn is_strong_blame(&self) -> bool

Provided Methods

fn is_strong_vote_for(&self, leader: AuthorityIndex) -> bool

True if this block is a strong vote pinned to leader.

fn is_strong_blame_for(&self, leader: AuthorityIndex) -> bool

True if this block is a strong blame pinned to leader.